Add common module that can be used to fetch externally facing IP address.
This commit adds a Terraform module that can be used to query an external
source in order to obtain externally visible IP address so that it can be
automatically added to the Security Group, etc. This is in lieu of lack of
such support in Terraform by default - there is no STUN protocol support,
or otherwise.

Signed-off-by: Krzysztof Wilczynski <kw@linux.com>