Fix. Update subnet-calculator module to work with Terraform 0.11.x, etc.
This commit fixes a problem where a variable without a splat has been referenced
in the output sourcing a value from a resources which is using the `count`
attribute. Since Terraform 0.11, the splat operator has to be explicitly used
when a resource has the `count` attribute - this is to avoid possible issues,
and thus it has been now enforced by Terraform.

Signed-off-by: Krzysztof Wilczynski <kw@linux.com>